+++ NodeJs/src/Views/VentaApp.js
@@ -3,6 +3,7 @@ import VentaService from "../../src/Controllers/VentaService";
 import CargaService from "../Controllers/CargaService";
 import CompradorService from "../Controllers/CompradorService";
 import { Link } from "react-router-dom";
+import Swal from 'sweetalert2';
 
 export default function VentaApp() {
   const [venta, setVenta] = useState([]);
@@ -71,17 +72,40 @@ export default function VentaApp() {
   }, [venta]);
 
   const deleteVenta = (id) => {
-    VentaService.delete(id)
-      .then((response) => {
-        listarVenta();
-      })
-      .catch((error) => {
-        console.log(error);
-      });
+    Swal.fire({
+      title: '¿Estás seguro de que deseas eliminar este registro?',
+      icon: 'warning',
+      showCancelButton: true,
+      confirmButtonText: 'Sí, eliminar',
+      cancelButtonText: 'Cancelar'
+    }).then((result) => {
+      if (result.isConfirmed) {
+        VentaService.delete(id)
+          .then((response) => {
+            listarVenta();
+            Swal.fire(
+              'Eliminado!',
+              'El registro ha sido eliminado.',
+              'success'
+            );
+          })
+          .catch((error) => {
+            console.log(error);
+            Swal.fire(
+              'Error!',
+              'Hubo un problema al eliminar el registro.',
+              'error'
+            );
+          });
+      }
+    });
   };
+
+
   const handleRowClick = (id) => {
     setselectedVenta(id);
   };
+
   useEffect(() => {
     const handleClickOutside = (event) => {
       if (tableRef.current && !tableRef.current.contains(event.target)) {
